+++ source/Tubeshade.Data/Migrations/000040_optional-duration.sql
@@ -1,0 +1,2 @@
+﻿ALTER TABLE media.videos
+    ALTER column duration DROP NOT NULL;