+++ source/Tubeshade.Server/Pages/Downloads/Index.cshtml.cs
@@ -9,7 +9,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using NodaTime;
 using Npgsql;
-using SponsorBlock;
 using Tubeshade.Data;
 using Tubeshade.Data.AccessControl;
 using Tubeshade.Data.Media;
@@ -167,21 +166,7 @@ public sealed class Index : PageModel, IDownloadPage, INonLibraryPage
         var videoIds = videos.Select(video => video.Id).ToArray();
         var segments = await _segmentRepository.GetForVideos(videoIds, userId, cancellationToken);
 
-        var models = videos.Select(video =>
-        {
-            var skippedDuration = segments
-                .Where(segment => segment.VideoId == video.Id && segment.Category != SegmentCategory.Filler)
-                .GetTotalDuration();
-
-            var actualDuration = (video.Duration - skippedDuration).Normalize();
-
-            return new VideoModel
-            {
-                Video = video,
-                ActualDuration = actualDuration,
-                Channel = Channels.Single(channel => video.ChannelId == channel.Id), // todo
-            };
-        }).ToList();
+        var models = videos.MapToModels(segments, Channels).ToList();
 
         var totalCount = videos is [] ? 0 : videos[0].TotalCount;