+++ source/Tubeshade.Server/Pages/Videos/VideoEntityExtensions.cs
@@ -1,6 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using JetBrains.Annotations;
+using SponsorBlock;
 using Tubeshade.Data.Media;
+using Tubeshade.Server.Pages.Shared;
 
 namespace Tubeshade.Server.Pages.Videos;
 
@@ -37,5 +42,36 @@ public static class VideoEntityExtensions
         return (attributes & FileAttributes.Directory) is FileAttributes.Directory
             ? Path.Combine(video.StoragePath, fileName)
             : Path.Combine(Path.GetDirectoryName(video.StoragePath) ?? string.Empty, fileName);
+    }
+
+    [LinqTunnel]
+    public static IEnumerable<VideoModel> MapToModels(
+        this IEnumerable<VideoEntity> videos,
+        List<SponsorBlockSegmentEntity> segments,
+        List<ChannelEntity> channels)
+    {
+        var channelDictionary = channels.ToDictionary(channel => channel.Id);
+        return videos.Select(video => video.MapToModel(segments, channelDictionary));
+    }
+
+    private static VideoModel MapToModel(
+        this VideoEntity video,
+        List<SponsorBlockSegmentEntity> segments,
+        Dictionary<Guid, ChannelEntity> channels)
+    {
+        var skippedDuration = segments
+            .Where(segment => segment.VideoId == video.Id && segment.Category != SegmentCategory.Filler)
+            .GetTotalDuration();
+
+        var actualDuration = video.Duration is { } duration
+            ? (duration - skippedDuration).Normalize()
+            : null;
+
+        return new VideoModel
+        {
+            Video = video,
+            ActualDuration = actualDuration,
+            Channel = channels[video.ChannelId], // todo: this probably should be done in SQL instead of in-memory
+        };
     }
 }