+++ source/Tubeshade.Server/Pages/Videos/VideoModel.cs
@@ -7,11 +7,11 @@ public sealed class VideoModel
 {
     public required VideoEntity Video { get; init; }
 
-    public required Period ActualDuration { get; init; }
+    public required Period? ActualDuration { get; init; }
 
     public required ChannelEntity Channel { get; init; }
 
     public bool Viewed => Video.Viewed is true;
 
-    public double DurationInSeconds => Video.Duration.ToDuration().TotalSeconds;
+    public double? DurationInSeconds => Video.Duration?.ToDuration().TotalSeconds;
 }